Render a scene of cubes
Add a render_scene method that will iterate through a list of instances in a scene. For now the scene is hardcoded and contains a list of instances, each with a model. The models are a couple of cubes. Add a method render_instance that will take the model, translate it to its desired position and render wireframe triangles between its vertices as described in the model's triangles.