Allow instance to receive an orientation_angle and scale arguments.
Use the position, orientation_angle and scale to obtain the corresponding transformation 4 x 4 matrices for translation, scale and rotation of the model in the scene. Apply the algorithm to find the resulting 4 x 4 matrix that represents the complete transformation. In the rasterizer, apply the transformation to each of the vertices by first making them into a vector of homogeneous coordinates (canonical) and the multiplying the resulting vector and the transformation matrix.