Implementing clipping of the objects in the scene.
As a first pass, discard objects that are completely behind the clipping planes. For each object, compare the signed distance from the center of the sphere enclosing the object to each clippling plane. If the signed distance is less than minus the sphere radius, it means the object is completely behind the clipping plane and should be discarded. We should not attempt to render this object at all to save in computational resources.