Clipping the objects in the scene, pass 2.
Clip all the triangles in each model in the scene against each of the clipping planes. Obtain the distance between the plane and each vertex of the triangle. If the distance is a positive number, the vertex is inside, if it's a negative number, the vertex is outside. If all the vertices are inside, the triangle is in and should be included among the clipped triangles. In any other case, we discard the triangle. For a better approximation, when only one vertex is in, we should create a new triangle between that vertex and the points where the sides of the triangle intersect the plane. Same for the case when two vertices are in, only in this case, the region of the triangle would be divided in two smaller triangles. This would change the collection of vertices in the model and the triangles. The images produced compare clipping with and without clipping triangles. In the second image, some triangles of the second model are missing because they would be only partially inside the bounding volume, so they are discarded.